swift_ride_backend/requirements/base.txt
# Base requirements
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.5
django-environ==0.11.2
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.6
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
Pillow==10.1.0
python-decouple==3.8
pyjwt==2.8.0
twilio==8.10.0
firebase-admin==6.2.0
google-cloud-storage==2.13.0
django-storages==1.14.2
django-phonenumber-field==7.2.0
phonenumbers==8.13.26
drf-yasg==1.21.7
django-redis==5.4.0
